FUCKING HELL! Look at this, no just look at this. Yet another reason why JavaScript is so fucking fucked up.
Not only is Date.parse not locale-aware, it actually messes up if you
give it a string that's not in that fucking ISO-weird-shit-YYYY-MM-DD
format.

For e.g. Change your locale to Indian English. Then say:

var fourthJuly = "4/7/2018"; // My PC uses the Indian locale for this
example.

// Fucking prints it as Sat Apr 07 2018 00:00:00 GMT+0530 (India Standard Time)
console.log(new Date(Date.parse(fourthJuly)).toString());

Next, do this. This is FUCKING WEIRD:

var thirtiethFebruary = "2/30/2018"; // Obviousy an invalid date as per
the Indian locale that I am using presently. I am
expecting a RangeError. But I would have expected that if the fucker did
actually value locales. Looks like it not only does not value them, it
first assumes this is still the US locale, God knows for what reason,
and then just eats any error and silently produces a unix timestamp for 2nd March
2018, i.e. 2 days after the last February date of the year.

// Produces Fri Mar 02 2018 00:00:00 GMT+0530 (India Standard Time)
console.log(new Date(Date.parse(thirtiethFebruary)).toString());

I have written an example in the date-app/date.js file. I am going to do
a video about it sometime. I first thought this was to do with the
browser locales, or the JS engine in Firefox, but the fucker produces
the same results in Node as well.