The following questions arose in my mind, to answer which, I made this example. They are:
Q1. What does the parent() property of a div read if it is absolutely positioned?
A1. The parent always refers to the parent element in the DOM, regardless of the position or the z-index properties.

Q2. And what if it has a separate layer of its own, i.e. it has a non-zero z-index?
A2. The answer to the previous bullet point answers this as well.

Q3. And in such a case, is its containing block the same as its parent?
A3. No, the containing block is not the same as the parent. As can be seen in the example above, #child1 and #child2, the two divs inside the #otherOuter div, have their parent as the #otherOuter, but they are positioned and sized as per the dimensions of the viewport, which becomes their containing block. Try sizing the window to different sizes and see their position and sizes change while the position and sizes of the other divs remain constant as they are relative to their containing blocks, which are their parents, because their immediate parents are also positioned absolutely.