drag and drop images using a file control (opacity hidden) with a label
In this example, I use a hidden file input control, i.e. an input control of type file, allowing the user to select only image files. Then, I display a preview of all the images the user has selected.

The only difference between this one and the previous one, 6.html, is that this one uses a label control instead of a button to allow the user to click on to select files. And since the label is for the file input control, the file input control's visibility should not be set to hidden using style = display: none; because then that input control won't receive events. Instead, it must be fooled into being invisible using any of the visual techniques such as the one set below (style = opacity: 0).