Updated the "drag and drop API" example in the MVC project 9 to do as follows:
In this example, I use the HTML5 drag and drop API to allow the user to drag and drop files into the form. When the user drops them, I filter out only the image files and display a preview for them.

The files in the dataTransfer.files object won't upload on their own when you post a form. So, on the button click of the submit button, I new up a FormData object and post it using an AJAX request made with jQuery. For more information about using FormData, please read the article on MDN titled Using FormData objects.