I made this example to answer this question for myself.
What happens if two HTML elements on my page happen to have the exact same ID and I have no liberty to change it? What's the default behavior of the DOM API in that case? Can I do something in that case to salvage the application?

I made a video about my observation here: https://youtu.be/ZHvn2BXlwdU