It appears that the keyup event is sent multiple times. That is, it fires many times.
The behaviour I observed was that it fired once the first time, then twice the second time, thrice the third time and so on.

I observed this while developing my website sathyaish.net.

I am doing this example to isolte the scenario and see if it is actually the case of multiple events firing or it is something with my code.

Solution
Nothing to do with jQuery. It's because the handlers need to be removed. They were being added again and again on the show functionality.