I wanted to find out if an expired cookie is sent to the server or not.
So, for e.g. there was a cookie named AUTH and it has now expired. Will the cookie remain in the Storage section in the Developer Tools window? Will it be sent by the client to the server with the expiration date set to the past date or will it just not be included in the HTTP payload? If I were writing the server side code, should I check for the absence of the cookie like this:

Request.Cookies["AUTH"] == null

or

Request.Cookies["AUTH"].Expires <= DateTime.Now

Also, if a cookie that has been deleted is sent to the server or not. If you're thinking, "Wait! A deleted cookie? Sent to the server? How can something that's been deleted be sent to the server? Won't it have to be re-created in order to be used if it was already deleted?"

No, no! You see, we know that when either the client or the server needs to delete a cookie

Already used cookies a thousand times. There are some things you remember and some things you'll forget. And I need to use them once again, so I want to test out a few things before I write the actual code. I am presently writing the Login functionality for my website https://sathyaish.net.